8/22 11:30 - 12:30 7章(P233 - 239) テキストデータを読み込めるようにしました。 床と穴のデータを反映させている途中。
