+++ PlasticRunner/Assets/Assets.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: dc46496950ad7514098c6dbf9f57681c
+folderAsset: yes
+timeCreated: 1440212472
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
