+++ PlasticRunner/Assets/Assets/Scripts/CameraControl.cs
