+++ PlasticRunner/Assets/Assets/Scripts/GameRoot.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1569d2f62056b66429004e3b582ec790
+timeCreated: 1440213340
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
