+++ PlasticRunner/Assets/Assets/Scripts/MapCreator.cs
@@ -12,6 +12,8 @@ public class Block {
 };
 
 public class MapCreator : MonoBehaviour {
+	private GameRoot game_root = null;
+	public TextAsset level_data_text = null;
 	public static float BLOCK_WIDTH = 1.0f;			// ブロックの幅
 	public static float BLOCK_HEIGHT = 0.2f;		// ブロックの高さ
 	public static int	BLOCK_NUM_IN_SCREEN = 24;	// 画面内に収まるブロックの個数
@@ -37,6 +39,8 @@ public class MapCreator : MonoBehaviour {
 
 		this.level_control = new LevelControl();
 		this.level_control.initialize();
+		this.level_control.loadLevelData(this.level_data_text);
+		this.game_root = this.gameObject.GetComponent<GameRoot>();
 	}
 
 	// Update is called once per frame
@@ -77,7 +81,8 @@ public class MapCreator : MonoBehaviour {
 		// これまでのコードで設定したblock_positionを渡す
 		//this.block_creator.createBlock(block_position);
 
-		this.level_control.update();	// LevelControlを更新
+		//this.level_control.update();	// LevelControlを更新
+		this.level_control.update(this.game_root.getPlayTime());
 
 		// level_controlに置かれたcurrent_block（今作るブロックの情報）の
 		// height（高さ）をシーン上の座標に変換
