+++ PlasticRunner/Assets/Assets/level_data.txt
@@ -1,0 +1,7 @@
+#				FLOOR			HOLE			HEIGHT
+# TIME	SPEED	MIN		MAX		MIN		MAX		MIN		MAX
+0.0		5.0		9		10		1		2		0		0
+15.0	6.0		8		9		3		5		0		0
+30.0	7.0		7		8		7		9		-4		4
+45.0	9.0		5		6		12		15		1		1
+60.0	8.0		6		7		10		12		2		2
