+++ sample_data/StudyProject/06_PlasticRunner_program/ProjectSettings/ProjectVersion.txt
@@ -1,0 +1,2 @@
+m_EditorVersion: 5.1.2f1
+m_StandardAssetsVersion: 0
